arcca-docker-lambda-reviews/image/requirements.txt
psycopg2-binary
selenium-wire==5.1.0
selenium==4.14.0
urllib3==2.0.6
certifi==2023.7.22
chromedriver-binary==120.0.6070.0.0
requests